requirements.txt
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
python-dotenv==1.2.1
Werkzeug==3.1.5
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48